pyproject.toml
[tool.poetry]
name = "nana"
version = "0.1.0"
description = ""
authors = ["Louis Anh Tran <louistrannamanh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
black = "^24.4.2"
passlib = "^1.7.4"
boto3 = "^1.34.141"
python-jose = "^3.3.0"
langchain = "^0.2.7"
openai = "^1.35.13"
pinecone-client = "^4.1.2"
pymupdf = "^1.24.7"
langchain-community = "^0.2.7"
langchain-openai = "^0.1.15"
pytest = "^8.3.2"
pandas = "^2.2.2"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
